Move the reward's durable close into the quest application module (#718)
The commit, its ambiguous-COMMIT reconciliation and the memory half that may
only follow a known COMMIT belong to the operation, not to the handler file
that plans the grants. Moving them takes `handlers/quest/rewards.rs` from 2232
to 2140 lines and tightens its ceiling to that figure; the file keeps its
remaining split obligation.

Two ceilings rise with their measurement: the composition root by the 6 lines
that install the durability port, and the quest test root by the 2-line
declaration of the reward-transaction scenario module.

Behavior is unchanged; only the owner and the visibility of the closing call.
